move spi files to server
