--- libs/core/src/main/java/org/elasticsearch/common/time/IsoCalendarDataProvider.java
@@ -1,23 +1,0 @@
-package org.elasticsearch.common.time;
-
-import java.util.Calendar;
-import java.util.Locale;
-import java.util.spi.CalendarDataProvider;
-
-public class IsoCalendarDataProvider extends CalendarDataProvider {
-
-    @Override
-    public int getFirstDayOfWeek(Locale locale) {
-        return Calendar.MONDAY;
-    }
-
-    @Override
-    public int getMinimalDaysInFirstWeek(Locale locale) {
-        return 4;
-    }
-
-    @Override
-    public Locale[] getAvailableLocales() {
-        return new Locale[]{Locale.ROOT};
-    }
-}
